feat(nx-cloud): set up nx workspace (#89) (#90)
* fix: upgrade solc from 0.7.6 to 0.8.29 (#74)

Snyk has created this PR to upgrade solc from 0.7.6 to 0.8.29.

See this package in npm:
solc

See this project in Snyk:
https://app.snyk.io/org/dargon789/project/36f3703d-86b2-4d1e-baeb-5644c446cffd?utm_source=github&utm_medium=referral&page=upgrade-pr



* feat: upgrade ethers from 5.8.0 to 6.13.6 (#73)

Snyk has created this PR to upgrade ethers from 5.8.0 to 6.13.6.

See this package in npm:
ethers

See this project in Snyk:
https://app.snyk.io/org/dargon789/project/36f3703d-86b2-4d1e-baeb-5644c446cffd?utm_source=github&utm_medium=referral&page=upgrade-pr



* fix: upgrade axios from 1.8.3 to 1.8.4 (#72)

Snyk has created this PR to upgrade axios from 1.8.3 to 1.8.4.

See this package in npm:
axios

See this project in Snyk:
https://app.snyk.io/org/dargon789/project/12f64623-bf9e-4ab1-9e99-ce83f08f7ff8?utm_source=github&utm_medium=referral&page=upgrade-pr



* feat: upgrade ethers from 5.8.0 to 6.13.6 (#76)

Snyk has created this PR to upgrade ethers from 5.8.0 to 6.13.6.

See this package in npm:
ethers

See this project in Snyk:
https://app.snyk.io/org/dargon789/project/72c16290-95f4-44ac-aec7-d00c3fb7e073?utm_source=github&utm_medium=referral&page=upgrade-pr



* Create jekyll-gh-pages.yml



* fix: upgrade @vscode/ripgrep from 1.15.6 to 1.15.11 (#78)

Snyk has created this PR to upgrade @vscode/ripgrep from 1.15.6 to 1.15.11.

See this package in yarn:
@vscode/ripgrep

See this project in Snyk:
https://app.snyk.io/org/dargon789/project/d9d81721-3222-4e91-88f5-136cc3c6b1ff?utm_source=github-cloud-app&utm_medium=referral&page=upgrade-pr




* fix: upgrade isomorphic-git from 1.24.2 to 1.30.1 (#77)

Snyk has created this PR to upgrade isomorphic-git from 1.24.2 to 1.30.1.

See this package in yarn:
isomorphic-git

See this project in Snyk:
https://app.snyk.io/org/dargon789/project/d9d81721-3222-4e91-88f5-136cc3c6b1ff?utm_source=github-cloud-app&utm_medium=referral&page=upgrade-pr




* feat: upgrade ethers from 5.8.0 to 6.13.6 (#82)

Snyk has created this PR to upgrade ethers from 5.8.0 to 6.13.6.

See this package in npm:
ethers

See this project in Snyk:
https://app.snyk.io/org/dargon789/project/4b4ed1a6-8e81-4e62-84b3-667f9a92e633?utm_source=github&utm_medium=referral&page=upgrade-pr



* fix: upgrade @remixproject/plugin-utils from 0.3.33 to 0.3.208 (#83)

Snyk has created this PR to upgrade @remixproject/plugin-utils from 0.3.33 to 0.3.208.

See this package in npm:
@remixproject/plugin-utils

See this project in Snyk:
https://app.snyk.io/org/dargon789/project/7c9a9f0f-9aed-4800-b8f4-a9fb3dc26e7d?utm_source=github&utm_medium=referral&page=upgrade-pr



---------

Signed-off-by: AU_gdev_19 <64915515+Dargon789@users.noreply.github.com>
Co-authored-by: snyk-bot <snyk-bot@snyk.io>
Co-authored-by: snyk-io[bot] <141718529+snyk-io[bot]@users.noreply.github.com>